backend/requirements.txt
Django
psycopg2

djangorestframework
markdown
django-filter

Pillow
ujson